# Performance improvements to iteration