Created a few more iterators to be used by the public interface for performance improvements. I did this in two commits, the second one changes the private interface signatures from funcs (technically same signature as ItemIterator) to Item, but since the public interface never actually exposes funcs for ascending I figured this would be okay.

Before:
```
~/.../github.com/cstockton/gbtree$ go test -test.bench=.*
1
PASS
BenchmarkInsert	 3000000	       411 ns/op
BenchmarkDelete	 3000000	       424 ns/op
BenchmarkGet	 5000000	       363 ns/op
BenchmarkIterateAscend	    3000	    474049 ns/op
BenchmarkIterateAscendLessThan	    5000	    274409 ns/op
BenchmarkIterateAscendGreaterOrEqual	    5000	    277208 ns/op
BenchmarkIterateAscendRange	   10000	    216873 ns/op
ok  	github.com/cstockton/gbtree	16.429s
```

After:
```
~/.../github.com/cstockton/btree$ go test -test.bench=.*
1
PASS
BenchmarkInsert	 3000000	       408 ns/op
BenchmarkDelete	 3000000	       422 ns/op
BenchmarkGet	 5000000	       360 ns/op
BenchmarkIterateAscend	    3000	    431754 ns/op
BenchmarkIterateAscendLessThan	    5000	    257723 ns/op
BenchmarkIterateAscendGreaterOrEqual	    5000	    257560 ns/op
BenchmarkIterateAscendRange	   10000	    205764 ns/op
ok  	github.com/cstockton/btree	15.984s
```