@gconnell 2015-06-22
When you get a chance, could you sign the Google CLA?  It's available at https://cla.developers.google.com/about/google-individual, and it needs to be signed before I accept code from you.  Basically, it says "you still own the code, but you allow us to use it".