@gconnell 2015-06-22
Thanks for signing!  I'll wait for your binary search changes before merging.